requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-core
langchain-pinecone
langchain-text-splitters
pinecone-client
sentence-transformers
groq
python-dotenv
pydantic
pydantic-settings
unstructured
pypdf
prometheus-client
mlflow
datasets
ragas
numpy
scikit-learn
httpx
markdown
tiktoken
